// frontend/src/api/model/githubCom4H1RZooraInternalDomainPlanPrice.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainBillingInterval } from './githubCom4H1RZooraInternalDomainBillingInterval';
import type { GithubCom4H1RZooraInternalDomainPlan } from './githubCom4H1RZooraInternalDomainPlan';

export interface GithubCom4H1RZooraInternalDomainPlanPrice {
  active?: boolean;
  /** minor units (Rial) */
  amount?: number;
  created_at?: string;
  currency?: string;
  id?: string;
  interval?: GithubCom4H1RZooraInternalDomainBillingInterval;
  plan?: GithubCom4H1RZooraInternalDomainPlan;
  updated_at?: string;
}