// frontend/src/api/model/githubCom4H1RZooraInternalDomainBillingInterval.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainBillingInterval = typeof GithubCom4H1RZooraInternalDomainBillingInterval[keyof typeof GithubCom4H1RZooraInternalDomainBillingInterval];


export const GithubCom4H1RZooraInternalDomainBillingInterval = {
  BillingIntervalMonthly: 'monthly',
  BillingIntervalYearly: 'yearly',
} as const;